Compare counts

Compare counts generated by counts()

Usage

compare_counts(
  df,
  interval = c("year", "month")[1],
  what = c("crashes", "fatalities", "injuries", "people")[1],
  where = list(states = "all", region = c("all", "ne", "mw", "s", "w")[1], urb = c("all",
    "rural", "urban")[1]),
  who = c("all", "drivers", "passengers", "bicyclists", "pedestrians")[1],
  involved = NULL,
  what2 = what,
  where2 = where,
  who2 = who,
  involved2 = involved
)

Arguments

df

The input FARS object.

interval

The interval in which to count: months or years.

what

What to count: crashes, fatalities, or people involved.

where

Where to count, a list with up to three elements: states ("all" by default), region ("all"), urb ("all")

who

The type of person to count: all (default) drivers, passengers, pedestrians, or bicyclists.

involved

Factors involved with the crash. Can be any of: distracted driver, drowsy driver, police pursuit, motorcycle, pedalcyclist, bicyclist, pedestrian, pedbike, young driver, older driver, speeding, alcohol, drugs, hit and run, roadway departure, rollover, or large trucks.

what2

Comparison point for 'what' (set to 'what' unless specified).

where2

Comparison point for 'where' (set to 'where' unless specified).

who2

Comparison point for 'who' (set to 'who' unless specified).

involved2

Comparison point for 'involved' (set to 'involved' unless specified).

Value

A tibble of counts.

Examples

# \donttest{
get_fars(years = 2020, states="Virginia") %>%
   compare_counts(
     where = list(urb="rural"),
     where2 = list(urb="urban")
     )
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpoNiPh3\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpoNiPh3\FARS data\raw\2020\format-64' already exists
#> # A tibble: 6 x 8
#> # Groups:   year [3]
#>    year date           n what    states region urb   who  
#>   <dbl> <date>     <int> <chr>   <chr>  <chr>  <chr> <chr>
#> 1  2019 2019-01-01   459 crashes all    all    rural all  
#> 2  2020 2020-01-01   408 crashes all    all    rural all  
#> 3  2021 2021-01-01   719 crashes all    all    rural all  
#> 4  2019 2019-01-01   312 crashes all    all    urban all  
#> 5  2020 2020-01-01   388 crashes all    all    urban all  
#> 6  2021 2021-01-01   815 crashes all    all    urban all  

get_fars(years = 2020, states="FL") %>%
   compare_counts(
     involved = "pedestrians",
     involved2 = "bicyclists"
     )
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpoNiPh3\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpoNiPh3\FARS data\raw\2020\format-64' already exists
<<<<<<< HEAD
#> # A tibble: 6 x 9
#> # Groups:   year [3]
=======
#> # A tibble: 2 × 9
#> # Groups:   year [1]
>>>>>>> aeabadf75a6c43010a374860b35a52c499ee72a8
#>    year date           n what    states region urb   who   involved   
#>   <dbl> <date>     <int> <chr>   <chr>  <chr>  <chr> <chr> <chr>      
#> 1  2019 2019-01-01   774 crashes all    all    all   all   pedestrians
#> 2  2020 2020-01-01  3097 crashes all    all    all   all   pedestrians
#> 3  2021 2021-01-01  1535 crashes all    all    all   all   pedestrians
#> 4  2019 2019-01-01   774 crashes all    all    all   all   bicyclists 
#> 5  2020 2020-01-01  3097 crashes all    all    all   all   bicyclists 
#> 6  2021 2021-01-01  1535 crashes all    all    all   all   bicyclists 
 # }