Generate counts

Use FARS or GES/CRSS data to generate commonly requested counts.

Usage

counts(
  df,
  what = c("crashes", "fatalities", "injuries", "people")[1],
  interval = c("year", "month")[1],
  where = list(states = "all", region = c("all", "ne", "mw", "s", "w")[1], urb = c("all",
    "rural", "urban")[1]),
  who = c("all", "drivers", "passengers", "bicyclists", "pedestrians")[1],
  involved = NULL,
  filterOnly = FALSE
)

Arguments

df

The input data object (must be of class 'FARS' or 'GESCRSS' as is produced by get_fars() and get_gescrss()).

what

What to count: crashes (the default), fatalities, injuries, or people involved.

interval

The interval in which to count: months or years (the default).

where

Where to count. Must be a list with any of the elements: states (can be 'all', full or abbreviated state names, or FIPS codes), region ('all', 'ne', 'mw', 's', or 'w'; short for northeast, midwest, south, and west), urb ('all', 'rural', or 'urban'). Any un-specified elements are set to 'all' by default.

who

The type of person to count: 'all' (default) 'drivers', 'passengers', 'pedestrians', or 'bicyclists'.

involved

Factors involved with the crash. Can be any of: 'distracted driver', 'drowsy driver', 'police pursuit', 'motorcycle', 'pedalcyclist', 'bicyclist', 'pedestrian', 'pedbike', 'young driver', 'older driver', 'speeding', 'alcohol', 'drugs', 'hit and run', 'roadway departure', 'rollover', or 'large trucks'. NULL by default.

filterOnly

Logical, whether to only filter data or reduce to counts (FALSE by default).

Value

Either a filtered tibble (filterOnly=TRUE) or a tibble of counts (filterOnly=FALSE). If filterOnly=TRUE, the tibble that is returned is the 'flat' tibble from the input FARS object, filtered according to other parameters.

If `df` is a GESCRSS object, the counts returned are the sum of the appropriate weights.

Examples

# \donttest{

get_fars(years = 2020) %>%
   counts(where = list(states="Virginia", urb="rural"))
<<<<<<< HEAD
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpqyJgwQ\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpqyJgwQ\FARS data\raw\2020\format-64' already exists
=======
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpioY24S\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpioY24S\FARS data\raw\2020\format-64' already exists
>>>>>>> a2d1be402faeb4330f2473d2714e271a18d058cc
#> # A tibble: 1 × 8
#> # Groups:   year [1]
#>    year date           n what    states   region urb   who  
#>   <dbl> <date>     <int> <chr>   <chr>    <chr>  <chr> <chr>
#> 1  2020 2020-01-01   408 crashes Virginia all    rural all  

get_fars(years = 2020) %>%
   counts(where=list(states="FL"), involved = "older driver")
<<<<<<< HEAD
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpqyJgwQ\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpqyJgwQ\FARS data\raw\2020\format-64' already exists
=======
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpioY24S\FARS data\raw\2020\format-32' already exists
#> Warning: 'C:\Users\s87ja\AppData\Local\Temp\RtmpioY24S\FARS data\raw\2020\format-64' already exists
>>>>>>> a2d1be402faeb4330f2473d2714e271a18d058cc
#> # A tibble: 1 × 9
#> # Groups:   year [1]
#>    year date           n what    states region urb   who   involved    
#>   <dbl> <date>     <int> <chr>   <chr>  <chr>  <chr> <chr> <chr>       
#> 1  2020 2020-01-01   625 crashes FL     all    all   all   older driver

# }