FARS Codebook

A table describing each FARS variable name, value, and corresponding value label.

Usage

fars_codebook

Format

A data frame with 15,951 rows and 19 variables:

source

The source of the data (either FARS or GES/CRSS).

file

The data file that contains the given variable.

name_ncsa

The original name of the data element.

name_rfars

The modified data element name used in rfars

label

The label of the data element itself (not its constituent values).

Definition

The data element's definition, pulled from the Analytical User Manual.

Additional Information

Additional information on the data element, pulled from the Analytical User Manual.

value

The original value of the data element.

value_label

The de-coded value label.

2014

Indicator: 1 if valid for 2014, NA otherwise.

2015

Indicator: 1 if valid for 2015, NA otherwise.

2016

Indicator: 1 if valid for 2016, NA otherwise.

2017

Indicator: 1 if valid for 2017, NA otherwise.

2018

Indicator: 1 if valid for 2018, NA otherwise.

2019

Indicator: 1 if valid for 2019, NA otherwise.

2020

Indicator: 1 if valid for 2020, NA otherwise.

2021

Indicator: 1 if valid for 2021, NA otherwise.

2022

Indicator: 1 if valid for 2022, NA otherwise.

2023

Indicator: 1 if valid for 2023, NA otherwise.

Details

This codebook serves as a useful reference for researchers using FARS data. The 'source' variable is intended to help combine with the gescrss_codebook. Data elements are relatively stable but are occasionally discontinued, created anew, or modified. The 'year' variable helps indicate the availability of data elements, and differentiates between different definitions over time. Users should always check for discontinuities when tabulating cases.

The 'file' variable indicates the file in which the given data element originally appeared. Here, files refers to the SAS files downloaded from NHTSA. Most data elements stayed in their original file. Those that did not were moved to the multi_ files. For example, 'weather' originates from the 'accident' file, but appears in the multi_acc data object created by rfars.

The 'name_ncsa' variable describes the data element's name as assigned by NCSA (the organization within NHTSA that manages the database). To maximize compatibility between years and ease of use for programming, 'name_rfars' provides a cleaned naming convention (via janitor::clean_names()).

Each data element has a 'label', a more human-readable version of the element names. For example, the label for 'road_fnc' is 'Roadway Function Class'. These are not definitions but may provide enough information to help users conduct their analysis. Consult the Analytical User’s Manual for definitions and further details.

'Definition' and 'Additional Information' were extracted from the Analytical User’s Manual.

Each data element has multiple 'value'-'value_label' pairs: 'value' represents the original, non-human-readable value (usually a number), and 'value_label' represents the corresponding text value. For example, for 'road_fnc', 1 (the 'value') corresponds to 'Rural-Principal Arterial-Interstate' (the 'value_label'), 2 corresponds to 'Rural-Principal Arterial-Other', etc.

@source Codebooks are automatically generated by extracting SAS format catalogs (.sas7bcat files) and VALUE statements from .sas files during data processing, then consolidating variable names, labels, and value-label mappings across all years into searchable reference tables. Source files are published by NHTSA and available here.

See also

Examples

head(rfars::fars_codebook)
#>    source     file name_ncsa name_rfars        label Definition
#>    <fctr>   <fctr>    <fctr>     <fctr>       <fctr>     <fctr>
#> 1:   FARS accident     STATE      state State Number       <NA>
#> 2:   FARS accident     STATE      state State Number       <NA>
#> 3:   FARS accident     STATE      state State Number       <NA>
#> 4:   FARS accident     STATE      state State Number       <NA>
#> 5:   FARS accident     STATE      state State Number       <NA>
#> 6:   FARS accident     STATE      state State Number       <NA>
#>                                                                                                                                                    Additional Information
#>                                                                                                                                                                    <fctr>
#> 1: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#> 2: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#> 3: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#> 4: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#> 5: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#> 6: GSA State data elements except for 43, Puerto Rico. The State in which the vehicle is registered, REG_STAT, is found in the Vehicle data file; the coding is the same.
#>     value value_label   2014   2015   2016   2017   2018   2019   2020   2021
#>    <fctr>      <fctr> <fctr> <fctr> <fctr> <fctr> <fctr> <fctr> <fctr> <fctr>
#> 1:      1     Alabama      1      1      1      1      1      1      1      1
#> 2:      2      Alaska      1      1      1      1      1      1      1      1
#> 3:      4     Arizona      1      1      1      1      1      1      1      1
#> 4:      5    Arkansas      1      1      1      1      1      1      1      1
#> 5:      6  California      1      1      1      1      1      1      1      1
#> 6:      8    Colorado      1      1      1      1      1      1      1      1
#>      2022   2023
#>    <fctr> <fctr>
#> 1:      1      1
#> 2:      1      1
#> 3:      1      1
#> 4:      1      1
#> 5:      1      1
#> 6:      1      1